remove a bunch of internal byte array extensions
update Buffers to indicate the 'toTypeArray' functions are little endian
