+++ core/src/commonMain/kotlin/com/littlekt/file/Buffers.kt
@@ -325,6 +325,9 @@ interface ByteBuffer : Buffer {
     }
 }
 
+/**
+ * Assumes the underlying [ByteArray] is in little endian order.
+ */
 fun ByteArray.put(value: Int, offset: Int): ByteArray {
     this[offset + 3] = (value shr 24).toByte()
     this[offset + 2] = (value shr 16).toByte()
@@ -333,6 +336,9 @@ fun ByteArray.put(value: Int, offset: Int): ByteArray {
     return this
 }
 
+/**
+ * Assumes the underlying [ByteArray] is in little endian order.
+ */
 fun ByteArray.put(value: Float, offset: Int): ByteArray {
     val bits = value.toRawBits()
     this[offset + 3] = (bits shr 24).toByte()
@@ -342,12 +348,18 @@ fun ByteArray.put(value: Float, offset: Int): ByteArray {
     return this
 }
 
+/**
+ * Assumes the underlying [ByteArray] is in little endian order.
+ */
 fun ByteArray.put(value: Short, offset: Int): ByteArray {
     this[offset + 1] = (value.toInt() shr 8).toByte()
     this[offset + 0] = value.toByte()
     return this
 }
 
+/**
+ * Converts the [FloatArray] to a little endian ordered [ByteArray].
+ */
 fun FloatArray.toByteArray(offset: Int = 0, size: Int = this.size): ByteArray {
     val floatArray = this
     val bytes = ByteArray(size * 4)
@@ -360,6 +372,9 @@ fun FloatArray.toByteArray(offset: Int = 0, size: Int = this.size): ByteArray {
     return bytes
 }
 
+/**
+ * Converts the [ByteArray] to a little endian [FloatArray].
+ */
 fun ByteArray.toFloatArray(): FloatArray {
     val byteArray = this
     val floats = FloatArray(byteArray.size / 4)
@@ -368,9 +383,9 @@ fun ByteArray.toFloatArray(): FloatArray {
         val byteI = i * 4
         val bits =
             (byteArray[byteI].toInt() and 0xFF) or
-                ((byteArray[byteI + 1].toInt() and 0xFF) shl 8) or
-                ((byteArray[byteI + 2].toInt() and 0xFF) shl 16) or
-                ((byteArray[byteI + 3].toInt() and 0xFF) shl 24)
+                    ((byteArray[byteI + 1].toInt() and 0xFF) shl 8) or
+                    ((byteArray[byteI + 2].toInt() and 0xFF) shl 16) or
+                    ((byteArray[byteI + 3].toInt() and 0xFF) shl 24)
 
         floats[i] = Float.fromBits(bits)
     }
@@ -378,6 +393,9 @@ fun ByteArray.toFloatArray(): FloatArray {
     return floats
 }
 
+/**
+ * Converts the [ByteArray] to a little endian [IntArray].
+ */
 fun ByteArray.toIntArray(): IntArray {
     val byteArray = this
     val ints = IntArray(byteArray.size / 4)
@@ -386,14 +404,17 @@ fun ByteArray.toIntArray(): IntArray {
         val byteI = i * 4
         ints[i] =
             byteArray[byteI].toInt() or
-                (byteArray[byteI + 1].toInt() shl 8) or
-                (byteArray[byteI + 2].toInt() shl 16) or
-                (byteArray[byteI + 3].toInt() shl 24)
+                    (byteArray[byteI + 1].toInt() shl 8) or
+                    (byteArray[byteI + 2].toInt() shl 16) or
+                    (byteArray[byteI + 3].toInt() shl 24)
     }
 
     return ints
 }
 
+/**
+ * Converts the [IntArray] to a little endian ordered [ByteArray].
+ */
 fun IntArray.toByteArray(): ByteArray {
     val intArray = this
     val bytes = ByteArray(intArray.size * 4)
@@ -405,6 +426,9 @@ fun IntArray.toByteArray(): ByteArray {
     return bytes
 }
 
+/**
+ * Converts the [ShortArray] to a little endian ordered [ByteArray].
+ */
 fun ShortArray.toByteArray(): ByteArray {
     val shortArray = this
     val bytes = ByteArray(shortArray.size * 2)
